feat/GLGDI: add uniform buffers to OpenGL GDI

+++ Source/Skyrocket/Core/Memory/Memory.hpp
@@ -19,7 +19,7 @@ namespace sky {
 /// @brief Contains a block of memory and its size
 struct MemoryBlock {
     /// @brief The byte size of the memory block
-    uint32_t size{0};
+    size_t size{0};
     /// @brief Pointer to the data being stored
     void* data{nullptr};
 };
